requirements.txt
APScheduler~=3.9.1.post1
python-dotenv~=0.21.1
fluentogram~=1.1.6
aiogram~=3.0.0b6
pydantic~=1.10.4
asyncpg~=0.27.0
aiohttp~=3.8.3